More debugging of parametric plot
Still getting different results for lambda against temperature when comparing to the previous solution for lambda when investigating the extra term. Almost as if extra temperature dependence is preventing any sort of change of phase, but things are a bit murky because of the lack of agreement. Also fixed K at zero temperature by not making binomial approximation. Parametric plot is not the main focus at the moment.